requirements.txt
requests==2.30.0
beautifulsoup4==4.12.2
aiogram==2.25.1
python-dotenv==1.0.0
lxml==4.9.2
